pyproject.toml
[project]
name = "ccal"
version = "0.1.0"
description = "CLI tool for adding calendar events"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "icalendar>=6.0",
    "keyring>=25.0",
    "litellm>=1.0",
    "pillow>=11.0",
    "pydantic>=2.13.1",
    "pytesseract>=0.3.13",
    "rich>=13.0",
    "typer>=0.24.1",
]

[project.scripts]
ccal = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
